+++ ichnaea/async/config.py
@@ -89,8 +89,6 @@ def configure_data(redis_client):
                                      queue_key='update_cellarea'),
         'update_cellarea_ocid': DataQueue('update_cellarea_ocid', redis_client,
                                           queue_key='update_cellarea_ocid'),
-        'update_mapstat': DataQueue('update_mapstat', redis_client,
-                                    queue_key='update_mapstat'),  # BBB
         'update_score': DataQueue('update_score', redis_client,
                                   queue_key='update_score'),
     }